x86/mtrr, pat: make pat independent from mtrr
Makes CONFIG_X86_PAT independent from CONFIG_MTRR by allowing pat to be
initialized even when MTRR is disabled/unsupported.
On some (virtual) platforms, PAT is supported without MTRR support
because PAT is designed to replace MTRR. In those environment, it's
desirable to allow PAT to be enabled even when MTRR is
disabled/unsupported.

With the change set of edfe63e, PAT is disabled when MTRR is disabled as
work around. Although the right solution is to allow PAT to be
initialized even when PAT is disabled, it was not addressed. This patch
addresses it.
This patch makes x86 pat indepdent from mtrr by splitting out common
rendezvous logic from MTRR code, and allows PAT to be initialized even
when MTRR is disabled/unsupported. CONFIG_X86_PAT_WITHOUT_MTRR as
default n is introduced to keep the old behavior for safety.

This patch also address the case of pat disabled.

Cc: Toshi Kani <toshi.kani@hp.com>
Cc: Borislav Petkov <bp@suse.de>
Link: http://lkml.kernel.org/r/1458769323-24491-6-git-send-email-toshi.kani@hpe.com
Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>