mini-sock: Guest device driver for mini sock device
Mini sock MMIO device is a virtual device for guest VM to communicate with
services running in VMM host.  It's an alternative to virtio-vsock.  Send
payload to the service in host and receive payload from the service in the
host.  Userspace VMM, e.g., qemu, maps the device to socket in the host.

The device accepts MMIO write of GPA of packet, and processes the packets
in order.  The packet carries a command and its payload.  Supported
commands are, REQUEST, RESPONSE, RST, SHUTDOWN, RW(SEND), RW(RECV).  The
following table shows the qemu mapping to unix datagram socket in host.

command         qemu behavior
------          -------------
REQUEST         socket() + bind() + connect() for unix dgram client socket
                socket() + bind() for unix dgram server socket
                socket() + bind() + connect() for stream client socket
                socket() + bind() + listen() for stream server socket
RESPONSE        accept() for stream server socket
SHUTDOWN        close()
RW(SEND)        send()
RW(RECV)        recv()
RST             reset device. repurposed this command for device, not for
                connection

The guest device driver provides API for userspace in guest to mini sock
device. The API is POST and COMPLETE.  POST is to queue the packet to the
device. COMPLETE is to wait for the command to complete.

The alternative is to map IOCTL to commands for mini sock device.  IOCTL of
CONNECT, LISTEN, ACCEPT, SHUTDOWN, SEND, RECV as high level bsd socket-like
API.  Implement POST/COMPLETE for the simplicity of the guest device
driver.  Those high level api can be implemented later on top of this in
kernel or small user library.

Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>