[REVERTME] x86/tdx: Enable extended topology for TD guest
Extended Topology is enumerated by CPUID leaf 0xb, and V2 Extended
Topology is enumerated by CPUID leaf 0x1f. The TD guest accessing the
leaf will cause #VE and return all zero, which means the (V2) Extended
topology is not supported.

In this case, it will limit the maximum number of addressable IDs for
logical processors, and the initial APIC ID is only 8-bit.

So, emulation of the leaf 0xb and 0x1f via hypercall are enabled to
support more logical processors for the TD guest.

Signed-off-by: Qian Wen <qian.wen@intel.com>