RFC KVM: guest_memfd: Add KVM arch callback for free folio
Make kswapd zero private pages before freeing to spread the cost instead of
the time of destroying the guest TD.

When destroying the guest TD, it zeros all guest memory with movdir64b().
If the guest memory size is hundreds of GB, it takes minutes.  Allow the
kswapd to do it to spread the cost with free_folio address space operation.

- Make KVM guest_memfd inode remain after releasing file and when all
  pages are freed, make inode evictable. Page cache for KVM guest can
  remain so that kswapd reclaims them.
- Add free_folio ops to the address space operation and call KVM arch
  callback.

The TDX KVM backend will register free folio callback to zero private pages
with movdir64b().

Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>