KVM: TDX: Allow parallel destruction of private KVM MMU page table
Allow parallel destruction of private KVM MMU page table.

It takes too long to destruct the TDX secure-EPT.  Tens of minute with
hundreds of GB guest memory.  This patch reduces the time by allowing
multiple vcpu threads to destruct it in parallel.

Create the task list of 1GB or root page tables, pick the one entry from
the list, and destruct of the page table entries under it.  Thus it avoids
multiple threads interfering with each other.  The destruction is split
into two phase, 1GB and root following the existing destruction logic.

The alternative option is to use PTE entry freezing to avoid race.  The
option is difficult to detect the completion because it requires to scan
the entries to confirm all entries are already processed.  It implies
all threads to do it.

Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>