KVM: x86: Pass is_private to gmem hook of gmem_max_level
TDX wants to know the faulting address is shared or private so that the max
level is limited by Secure-EPT or not.  Because fault->gfn doesn't include
shared bit, gfn doesn't tell if the faulting address is shared or not.
Pass is_private for TDX case.

TDX logic will be if (!is_private) return 0; else return PG_LEVEL_4K.

Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>