KVM: x86: Add a hook in kvm_arch_vcpu_map_memory()
In the case of SEV and TDX, the memory contents needs to be provided to be
encrypted when populating guest memory before running the guest.  Add a
hook in kvm_mmu_map_tdp_page() for KVM_MAP_MEMORY before calling
kvm_mmu_tdp_page() to adjust error code for page fault.  SEV and TDX KVM
will use the hooks.

Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>
---
v2:
- Make pre_mmu_map_page() to take error_code.
- Drop post_mmu_map_page().
- Drop struct kvm_memory_map.source check.