KVM: selftests: TDX: Test case to check the TSC offset is correct
TDX module manages the TSC offset.  The VMM can't change it.  KVM also
maintains the TSC offset and updates it on vCPU creation and other
occasions for the VMX guest.  KVM prohibits updating the TSC offset for the
TDX guest.  (or vcpu->arch.guest_tsc_protected is true.)

Test it by checking if the TSC offset of multiple vCPUs is the same.  It is
a compromise to compare the TSC offset of two vCPUs because it's difficult
for the user space program to get the TSC offset value in the TDX module.
The right thing to do is to check if the TSC offset of KVM is the same as
the TSC OFFSET of VMCS metadata.

Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>